// booking-ui/src/api/types.ts
export interface paths {
    "/api/admin/bookings": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["AdminBookings_listUpcomingBookings"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/admin/bookings/{id}/cancel": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["AdminBookings_cancelBooking"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/admin/event-types": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["AdminEventTypes_listEventTypes"];
        put?: never;
        post: operations["AdminEventTypes_createEventType"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/admin/event-types/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put: operations["AdminEventTypes_updateEventType"];
        post?: never;
        delete: operations["AdminEventTypes_deleteEventType"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/bookings": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["PublicBookings_createBooking"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/bookings/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["PublicBookings_getBooking"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/bookings/{id}/cancel": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["PublicBookings_cancelBooking"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/events": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["PublicEvents_listEventTypes"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/events/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["PublicEvents_getEventType"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/events/{id}/slots": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["PublicEvents_getSlots"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        Booking: {
            id: string;
            eventTypeId: string;
            slotStartTime: string;
            slotEndTime: string;
            guestName: string;
            guestEmail: string;
            status: components["schemas"]["BookingStatus"];
            createdAt: string;
            cancelledAt: string | null;
            cancelReason: string | null;
        };
        /** @enum {string} */
        BookingStatus: "confirmed" | "cancelled";
        CancelBookingRequest: {
            cancelReason: string | null;
        };
        CreateBookingRequest: {
            eventTypeId: string;
            slotStartTime: string;
            guestName: string;
            guestEmail: string;
        };
        CreateEventTypeRequest: {
            name: string;
            description: string | null;
            /** Format: int32 */
            durationMinutes: number;
        };
        ErrorResponse: {
            code: string;
            message: string;
            details: string | null;
        };
        EventType: {
            id: string;
            name: string;
            description: string | null;
            /** Format: int32 */
            durationMinutes: number;
            isActive: boolean;
            createdAt: string;
            updatedAt: string;
        };
        Slot: {
            startTime: string;
            endTime: string;
            isAvailable: boolean;
        };
        UpdateEventTypeRequest: {
            name: string;
            description: string | null;
            /** Format: int32 */
            durationMinutes: number;
            isActive: boolean;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}